[llvm/llvm-project][Coroutines]  Move the SuspendCrossingInfo analysis helper into its own header/source
* Move SuspendCrossingInfo to its own files to clean up CoroFrame
